refactor(naming): move tests to separate directory and rename files to match snake case consistently
