+++ tests/test_azure_endpoints.py
